Finished implementation of Resource class Update documentation Uodate tests
