+++ README.md
@@ -58,18 +58,23 @@ model.then(function(schema) {
 
 Following methods are available on `Schema` instances:
 
+* `cast(fieldName, value, index, skipConstraints)` - returns a value cast against a named `fieldName`
+* `test(fieldName, value, index, skipConstraints)` - returns boolean after a check if value can be casted against a named `fieldName`
+* `convertRow(items, failFast = false, skipConstraints = false)` - convert the arguments given to the types of the current schema <sup>1</sup>
+* `convert(items, failFast = false, skipConstraints = false)` - convert an array of rows using the current schema of the Schema instance <sup>1</sup>
 * `fields` - returns an array of the schema's fields
 * `foreignKeys` - returns the foreign key property for the schema
-* `getConstraints(fieldName, index = 0)` - return the constraints object for a given `fieldName` <sup>1</sup>
-* `getField(fieldName, index = 0)` - return the field object for `fieldName` <sup>1</sup>
+* `getConstraints(fieldName, index = 0)` - return the constraints object for a given `fieldName` <sup>2</sup>
+* `getField(fieldName, index = 0)` - return the field object for `fieldName` <sup>2</sup>
 * `getFieldsByType(typeName)` - return all fields that match the given type
+* `getRequiredHeaders` - returns headers with the `required` constraint as an array
+* `getUniqueHeaders` - returns headers with the `unique` constraint as an array
 * `hasField(fieldName)` - checks if the field exists in the schema. Returns a boolean
 * `headers` - returns an array of the schema headers
-* `primaryKey` - returns the primary key field for the schema
-* `requiredHeaders` - returns headers with the `required` constraint as an array
-* `uniqueHeaders` - returns headers with the `unique` constraint as an array
-  
-<sup>1</sup> Where the optional index argument is available, it can be used as a positional argument if the schema has multiple fields with the same name
+* `primaryKey` - returns the primary key field for the schema  
+
+<sup>1</sup> Where the option `failFast` is given, it will raise the first error it encounters, otherwise an array of errors thrown (if there are any errors occur)  
+<sup>2</sup> Where the optional index argument is available, it can be used as a positional argument if the schema has multiple fields with the same name
 
 ### Infer
 
@@ -310,30 +315,28 @@ Available types, formats and resultant value of the cast:
 <sup>4</sup> topojson is not implemented
 
 ### Resource
+
+instance always returns `Promise`
+
 ```javascript
 var parse = require('csv-parse');
 var fs = require('fs');
-var infer = require('jsontableschema').infer;
-var Schema = require('jsontableschema').schema;
+var Infer = require('jsontableschema').infer;
 var Resource = require('jsontableschema').resource;
 
 fs.readFile('/path/to/example.csv', function(err, data) {
   parse(data, function(error, values) {
     var headers = values.shift()
-        , schema = infer(headers, values)
-        , model = new Schema(schema)
+        , schema = Infer(headers, values)
+        , model = new Resource(schema, values)
         
-    model.then(function(schema) {
-        // working code to use schema model
-        var resource = new Resource(schema, values)
-        
+    model.then(function(resource) {
         try {
-            var converted = resource.convert(values)
-            // do something with converted rows
+            var converted = resource.iter()
+            // do something with converted data
         } catch(e) {
           // conversion failed
         }
-       
     }).catch(function(error) {
         // something went wrong and error variable has explanations
     })
@@ -341,14 +344,8 @@ fs.readFile('/path/to/example.csv', function(err, data) {
 });
 ```
 
-Following methods are available on `Schema` instances:
-
-* `cast(fieldName, value, index, skipConstraints)` - returns a value cast against a named `fieldName`
-* `test(fieldName, value, index, skipConstraints)` - returns boolean after a check if value can be casted against a named `fieldName`
-* `convertRow(items, failFast = false, skipConstraints = false)` - convert the arguments given to the types of the current schema <sup>1</sup>
-* `convert(items, failFast = false, skipConstraints = false)` - convert an array of rows using the current schema of the Schema instance <sup>1</sup>
-
-<sup>1</sup> Where the option `failFast` is given, it will raise the first error it encounters, otherwise an array of errors thrown (if there are any errors occur)
+Following methods are available on `Resource` instances:
+* `iter(failFast = false, skipConstraints = false)` - Iter through the given dataset and return the converted dataset
 
 ## Goals
