+++ src/schema.js
@@ -3,6 +3,8 @@ import 'isomorphic-fetch'
 import url from 'url'
 import validate from './validate'
 import utilities from './utilities'
+import Type from './types'
+import constraints from './constraints'
 
 /**
  * Model for a JSON Table Schema.
@@ -22,7 +24,129 @@ import utilities from './utilities'
 export default class Schema {
   constructor(source, caseInsensitiveHeaders = false) {
     this.caseInsensitiveHeaders = !!caseInsensitiveHeaders
+    this.type = new Type()
     return load(this, source)
+  }
+
+  /**
+   * Cast value to fieldName's type
+   *
+   * @param fieldName
+   * @param value
+   * @param index
+   * @param skipConstraints
+   *
+   * @returns {Type}
+   * @throws Error if value can't be casted
+   */
+  cast(fieldName, value, index = 0, skipConstraints = true) {
+    const field = this.getField(fieldName, index)
+    return this.type.cast(field, value, skipConstraints)
+  }
+
+  /**
+   * Check if value to fieldName's type can be casted
+   *
+   * @param fieldName
+   * @param value
+   * @param index
+   * @param skipConstraints
+   *
+   * @returns {Boolean}
+   */
+  test(fieldName, value, index = 0, skipConstraints = true) {
+    const field = this.getField(fieldName, index)
+    return this.type.test(field, value, skipConstraints)
+  }
+
+  /**
+   * Convert the arguments given to the types of the current schema. Last
+   * argument could be { failFast: true|false }.  If the option `failFast` is
+   * given, it will raise the first error it encounters, otherwise an array of
+   * errors thrown (if there are any errors occur)
+   *
+   * @param items
+   * @param failFast
+   * @param skipConstraints
+   * @returns {Array}
+   */
+  convertRow(items, failFast = false, skipConstraints = false) {
+    const headers = this.headers()
+      , result = []
+      , errors = []
+
+    if (headers.length !== items.length) {
+      throw new Error('The number of items to convert does not match the ' +
+                      'number of fields given in the schema')
+    }
+
+    for (let i = 0, length = items.length; i < length; i++) {
+      try {
+        const fieldName = headers[i]
+          , value = this.cast(fieldName, items[i], i, skipConstraints)
+
+        if (!skipConstraints) {
+          // unique constraints available only from Resource
+          if (this.uniqueness && this.uniqueHeaders) {
+            constraints.check_unique(fieldName, value, this.uniqueHeaders,
+                                     this.uniqueness)
+          }
+        }
+        result.push(value)
+      } catch (e) {
+        let error
+        switch (e.name) {
+          case 'UniqueConstraintsError':
+            error = e.message
+            break
+          default:
+            error =
+              `Wrong type for header: ${headers[i]} and value: ${items[i]}`
+        }
+        if (failFast === true) {
+          throw new Error(error)
+        } else {
+          errors.push(error)
+        }
+      }
+    }
+
+    if (errors.length > 0) {
+      throw errors
+    }
+    return result
+  }
+
+  /**
+   * Convert an array of rows to the types of the current schema. If the option
+   * `failFast` is given, it will raise the first error it encounters,
+   * otherwise an array of errors thrown (if there are any errors occur)
+   *
+   * @param items
+   * @param failFast
+   * @param skipConstraints
+   * @returns {Array}
+   */
+  convert(items, failFast = false, skipConstraints = false) {
+    const result = []
+    let errors = []
+
+    for (const item of items) {
+      try {
+        result.push(this.convertRow(item, failFast, skipConstraints))
+      } catch (e) {
+        if (failFast === true) {
+          throw e
+        } else {
+          errors = errors.concat(e)
+        }
+      }
+    }
+
+    if (errors.length > 0) {
+      throw errors
+    }
+    return result
   }
 
   /**
@@ -91,6 +215,28 @@ export default class Schema {
   }
 
   /**
+   * Get all headers with required constraints set to true
+   * @returns {Array}
+   */
+  getRequiredHeaders() {
+    return _.chain(this.descriptor.fields)
+      .filter(field => field.constraints.required === true)
+      .map(field => field.name)
+      .value()
+  }
+
+  /**
+   * Get all headers with unique constraints set to true
+   * @returns {Array}
+   */
+  getUniqueHeaders() {
+    return _.chain(this.descriptor.fields)
+      .filter(field => field.constraints.unique === true)
+      .map(field => field.name)
+      .value()
+  }
+
+  /**
    * Check if the field exists in the schema
    *
    * @param fieldName
@@ -119,28 +265,6 @@ export default class Schema {
    */
   primaryKey() {
     return this.descriptor.primaryKey
-  }
-
-  /**
-   * Get all headers with required constraints set to true
-   * @returns {Array}
-   */
-  requiredHeaders() {
-    return _.chain(this.descriptor.fields)
-      .filter(field => field.constraints.required === true)
-      .map(field => field.name)
-      .value()
-  }
-
-  /**
-   * Get all headers with unique constraints set to true
-   * @returns {Array}
-   */
-  uniqueHeaders() {
-    return _.chain(this.descriptor.fields)
-      .filter(field => field.constraints.unique === true)
-      .map(field => field.name)
-      .value()
   }
 }
