+++ test/resource.js
@@ -39,205 +39,18 @@ describe('Resource', () => {
     done()
   })
 
-  it('should return true on test', done => {
-    (new Schema(SCHEMA)).then(schema => {
-      const resource = new Resource(schema)
-      assert.isTrue(resource.test('age', 1))
-      done()
-    }, (error) => {
-      assert.isNull(error)
-      done()
-    })
-  })
-
-  it('convert row', done => {
-    (new Schema(SCHEMA)).then(schema => {
-      const resource = new Resource(schema)
-        , value = ['string', '10.0', '1', 'string', 'string']
-        , convertedRow = resource.convertRow(value)
-      assert.deepEqual(['string', '10.0', 1, 'string', 'string'], convertedRow)
-      done()
-    }, error => {
-      assert.isNull(error)
-      done()
-    })
-  })
-
-  it('shouldn\'t convert row with less items than headers count', done => {
-    (new Schema(SCHEMA)).then(schema => {
-      const resource = new Resource(schema)
-      assert.throws(() => {
-        resource.convertRow(['string', '10.0', '1', 'string'])
-      }, Error)
-      done()
-    }, error => {
-      assert.isNull(error)
-      done()
-    })
-  })
-
-  it('shouldn\'t convert row with too many items', done => {
-    (new Schema(SCHEMA)).then(schema => {
-      const resource = new Resource(schema)
-      assert.throws(() => {
-        resource.convertRow(
-          ['string', '10.0', '1', 'string', 'string', 'string'])
-      }, Error)
-      done()
-    }, error => {
-      assert.isNull(error)
-      done()
-    })
-  })
-
-  it('shouldn\'t convert row with wrong type (fail fast)', done => {
-    (new Schema(SCHEMA)).then(schema => {
-      const resource = new Resource(schema)
-      assert.throws(() => {
-        resource.convertRow(['string', 'notdecimal', '10.6', 'string', 'string']
-          , true)
-      }, Error)
-      done()
-    }, error => {
-      assert.isNull(error)
-      done()
-    })
-  })
-
-  it('shouldn\'t convert row with wrong type multiple errors', done => {
-    (new Schema(SCHEMA)).then(schema => {
-      const resource = new Resource(schema)
-      try {
-        resource.convertRow(['string', 'notdecimal', '10.6', 'string', true])
-      } catch (e) {
-        assert.isArray(e)
-        assert.equal(e.length, 3)
-      }
-      done()
-    }, error => {
-      assert.isNull(error)
-      done()
-    })
-  })
-
-  it('convert multiple rows', done => {
-    (new Schema(SCHEMA)).then(schema => {
-      const resource = new Resource(schema)
-        , rows = resource.convert(
-        [
-          ['string', '10.0', '1', 'string', 'string']
-          , ['string', '10.0', '1', 'string', 'string']
-          , ['string', '10.0', '1', 'string', 'string']
-          , ['string', '10.0', '1', 'string', 'string']
-          , ['string', '10.0', '1', 'string', 'string']
-        ])
-      for (const row of rows) {
-        assert.deepEqual(['string', '10.0', 1, 'string', 'string'], row)
-      }
-      done()
-    }, error => {
-      assert.isNull(error)
-      done()
-    })
-  })
-
-  it('shouldn\'t convert rows with wrong types fail fast', done => {
-    (new Schema(SCHEMA)).then(schema => {
-      const resource = new Resource(schema)
-      assert.throws(() => {
-        resource.convert(
-          [
-            ['string', 'not', '1', 'string', 'string']
-            , ['string', '10.0', '1', 'string', 'string']
-            , ['string', 'an', '1', 'string', 'string']
-            , ['string', '10.0', '1', 'string', 'string']
-            , ['string', '10.0', 'integer', 'string', 'string']
-          ], true)
-      }, Error)
-      done()
-    }, error => {
-      assert.isNull(error)
-      done()
-    })
-  })
-
-  it('shouldn\'t convert rows with wrong types multiple errors', done => {
-    (new Schema(SCHEMA)).then(schema => {
-      const resource = new Resource(schema)
-      try {
-        resource.convert([['string', 'not', '1', 'string', 'string']
-                           , ['string', '10.0', '1', 'string', 'string']
-                           , ['string', 'an', '1', 'string', 'string']
-                           , ['string', '10.0', '1', 'string', 'string']
-                           , ['string', '10.0', 'integer', 'string', 'string']])
-      } catch (e) {
-        assert.isArray(e)
-        assert.equal(e.length, 3)
-      }
-      done()
-    }, error => {
-      assert.isNull(error)
-      done()
-    })
-  })
-
-  it('shouldn\'t convert rows with incorrect row length fail fast', done => {
-    (new Schema(SCHEMA)).then(schema => {
-      const resource = new Resource(schema)
-      assert.throws(() => {
-        resource.convert(
-          [
-            ['string', '10.0', '1', 'string']
-            , ['string', '10.0', '1', 'string', 'string']
-            , ['string', '10.0', '1', 'string', 'string', 1]
-            , ['string', '10.0', '1', 'string', 'string']
-            , ['string', '10.0', '1', 'string', 'string']
-          ], true)
-      }, Error)
-      done()
-    }, error => {
-      assert.isNull(error)
-      done()
-    })
-  })
-
-  it('shouldn\'t convert rows with incorrect row length multiple errors',
-     done => {
-       (new Schema(SCHEMA)).then(schema => {
-         const resource = new Resource(schema)
-         try {
-           resource.convert(
-             [['string', '10.0', '1', 'string']
-               , ['string', '10.0', '1', 'string', 'string']
-               , ['string', '10.0', '1', 'string', 'string', 1]
-               , ['string', '10.0', '1', 'string', 'string']
-               , ['string', '10.0', '1', 'string', 'string']])
-         } catch (e) {
-           assert.isArray(e)
-           assert.equal(e.length, 2)
-         }
-         done()
-       }, error => {
-         assert.isNull(error)
-         done()
-       })
-     })
-
   it('unique constraints violation', done => {
     SCHEMA.fields[0].constraints.unique = true
-
-    const model = new Schema(SCHEMA)
-    model.then(schema => {
-      const resource = new Resource(schema)
+    const data = [
+      ['string', '10.0', '1', 'string', 'string']
+      , ['string2', '10.0', '1', 'string', 'string']
+      , ['string3', '10.0', '1', 'string', 'string']
+      , ['string4', '10.0', '1', 'string', 'string']
+      , ['string', '10.0', '1', 'string', 'string']
+    ];
+    (new Resource(SCHEMA, data)).then(resource => {
       assert.throws(() => {
-        resource.convert(
-          [
-            ['string', '10.0', '1', 'string', 'string']
-            , ['string2', '10.0', '1', 'string', 'string']
-            , ['string3', '10.0', '1', 'string', 'string']
-            , ['string4', '10.0', '1', 'string', 'string']
-            , ['string', '10.0', '1', 'string', 'string']
-          ])
+        resource.iter()
       }, Array)
       done()
     }, error => {
