+++ test/schema.js
@@ -61,7 +61,7 @@ describe('Models', () => {
 
   it('have a correct number of header required columns', done => {
     (new Schema(SCHEMA, true)).then(schema => {
-      const headers = schema.requiredHeaders()
+      const headers = schema.getRequiredHeaders()
       assert.equal(headers.length, 2)
       assert.equal(headers[0], 'id')
       assert.equal(headers[1], 'name')
@@ -181,8 +181,8 @@ describe('Models', () => {
       , model = new Schema(data)
 
     model.then(schema => {
-      assert.isArray(schema.requiredHeaders())
-      assert.equal(schema.requiredHeaders().length, 1)
+      assert.isArray(schema.getRequiredHeaders())
+      assert.equal(schema.getRequiredHeaders().length, 1)
       done()
     }, (error) => {
       assert.isNull(error)
@@ -225,7 +225,7 @@ describe('Models', () => {
     const model = new Schema(url, true)
     model.then((schema) => {
       assert.equal(schema.headers().length, 5)
-      assert.equal(schema.requiredHeaders().length, 2)
+      assert.equal(schema.getRequiredHeaders().length, 2)
       assert.isTrue(schema.hasField('id'))
       assert.isTrue(schema.hasField('height'))
       assert.isTrue(schema.hasField('age'))
@@ -252,4 +252,177 @@ describe('Models', () => {
       done()
     })
   })
+
+  it('should return true on test', done => {
+    (new Schema(SCHEMA)).then(schema => {
+      assert.isTrue(schema.test('height', 1))
+      done()
+    }, (error) => {
+      assert.isNull(error)
+      done()
+    })
+  })
+
+  it('convert row', done => {
+    (new Schema(SCHEMA)).then(schema => {
+      const value = ['string', '10.0', '1', 'string', 'string']
+        , convertedRow = schema.convertRow(value)
+      assert.deepEqual(['string', '10.0', 1, 'string', 'string'], convertedRow)
+      done()
+    }, error => {
+      assert.isNull(error)
+      done()
+    })
+  })
+
+  it('shouldn\'t convert row with less items than headers count', done => {
+    (new Schema(SCHEMA)).then(schema => {
+      assert.throws(() => {
+        schema.convertRow(['string', '10.0', '1', 'string'])
+      }, Error)
+      done()
+    }, error => {
+      assert.isNull(error)
+      done()
+    })
+  })
+
+  it('shouldn\'t convert row with too many items', done => {
+    (new Schema(SCHEMA)).then(schema => {
+      assert.throws(() => {
+        schema.convertRow(
+          ['string', '10.0', '1', 'string', 'string', 'string'])
+      }, Error)
+      done()
+    }, error => {
+      assert.isNull(error)
+      done()
+    })
+  })
+
+  it('shouldn\'t convert row with wrong type (fail fast)', done => {
+    (new Schema(SCHEMA)).then(schema => {
+      assert.throws(() => {
+        schema.convertRow(['string', 'notdecimal', '10.6', 'string', 'string']
+          , true)
+      }, Error)
+      done()
+    }, error => {
+      assert.isNull(error)
+      done()
+    })
+  })
+
+  it('shouldn\'t convert row with wrong type multiple errors', done => {
+    (new Schema(SCHEMA)).then(schema => {
+      try {
+        schema.convertRow(['string', 'notdecimal', '10.6', 'string', true])
+      } catch (e) {
+        assert.isArray(e)
+        assert.equal(e.length, 3)
+      }
+      done()
+    }, error => {
+      assert.isNull(error)
+      done()
+    })
+  })
+
+  it('convert multiple rows', done => {
+    (new Schema(SCHEMA)).then(schema => {
+      const rows = schema.convert(
+        [
+          ['string', '10.0', '1', 'string', 'string']
+          , ['string', '10.0', '1', 'string', 'string']
+          , ['string', '10.0', '1', 'string', 'string']
+          , ['string', '10.0', '1', 'string', 'string']
+          , ['string', '10.0', '1', 'string', 'string']
+        ])
+      for (const row of rows) {
+        assert.deepEqual(['string', '10.0', 1, 'string', 'string'], row)
+      }
+      done()
+    }, error => {
+      assert.isNull(error)
+      done()
+    })
+  })
+
+  it('shouldn\'t convert rows with wrong types fail fast', done => {
+    (new Schema(SCHEMA)).then(schema => {
+      assert.throws(() => {
+        schema.convert(
+          [
+            ['string', 'not', '1', 'string', 'string']
+            , ['string', '10.0', '1', 'string', 'string']
+            , ['string', 'an', '1', 'string', 'string']
+            , ['string', '10.0', '1', 'string', 'string']
+            , ['string', '10.0', 'integer', 'string', 'string']
+          ], true)
+      }, Error)
+      done()
+    }, error => {
+      assert.isNull(error)
+      done()
+    })
+  })
+
+  it('shouldn\'t convert rows with wrong types multiple errors', done => {
+    (new Schema(SCHEMA)).then(schema => {
+      try {
+        schema.convert([['string', 'not', '1', 'string', 'string']
+                         , ['string', '10.0', '1', 'string', 'string']
+                         , ['string', 'an', '1', 'string', 'string']
+                         , ['string', '10.0', '1', 'string', 'string']
+                         , ['string', '10.0', 'integer', 'string', 'string']])
+      } catch (e) {
+        assert.isArray(e)
+        assert.equal(e.length, 3)
+      }
+      done()
+    }, error => {
+      assert.isNull(error)
+      done()
+    })
+  })
+
+  it('shouldn\'t convert rows with incorrect row length fail fast', done => {
+    (new Schema(SCHEMA)).then(schema => {
+      assert.throws(() => {
+        schema.convert(
+          [
+            ['string', '10.0', '1', 'string']
+            , ['string', '10.0', '1', 'string', 'string']
+            , ['string', '10.0', '1', 'string', 'string', 1]
+            , ['string', '10.0', '1', 'string', 'string']
+            , ['string', '10.0', '1', 'string', 'string']
+          ], true)
+      }, Error)
+      done()
+    }, error => {
+      assert.isNull(error)
+      done()
+    })
+  })
+
+  it('shouldn\'t convert rows with incorrect row length multiple errors',
+     done => {
+       (new Schema(SCHEMA)).then(schema => {
+         try {
+           schema.convert(
+             [['string', '10.0', '1', 'string']
+               , ['string', '10.0', '1', 'string', 'string']
+               , ['string', '10.0', '1', 'string', 'string', 1]
+               , ['string', '10.0', '1', 'string', 'string']
+               , ['string', '10.0', '1', 'string', 'string']])
+         } catch (e) {
+           assert.isArray(e)
+           assert.equal(e.length, 2)
+         }
+         done()
+       }, error => {
+         assert.isNull(error)
+         done()
+       })
+     })
 })
